--- test/scenarios/default.py
@@ -1,5 +1,0 @@
-def prestop():
-    pass
-
-
-run_tests()
